Function changes
Changed the placeCar function so we are now able to grab it. Change the print function as well. Now it highlights the cursor on the board. I want to add a full piece highlight when i select a piece and i want the cursor to move witht he selected index and be placed on the placed index and not its originla selected location