Game logic
Fixed the issue where the game would break when the red car piece would move out of bounds from the map. Logic was fine in the function but it was logic in the draw preview and erase that was messing it up. added a guard to it and returned imiedialty after the won the game in the main. Will start to work on the mas and how to implament levels so as you go they get harder. or you can pick, maybe both