Add color support and added levels
Added functions to change colors of the pieces on the board. Also changed the game logic to allow us to reset the games once we finish all maps. Pretty much done with the game. Will add a held counter where it will count how many times user pick up a piece and maybe a timer per each map. Will try to store this in a hashmap so user can see previous atemps and times