Notes and stop watch implamentaion
First created a move counter so when user picks up a piece it will incrament the total amount of moves user has done for the current level.

Started to add a stop watch timer, works fine for the first round but then breaks the game when i move onto a next level. need to fix this issue